feat(mesure): refactor mesure actions
