+++ packages/api/src/routes/hasura-actions/mesure/actions/index.js
@@ -1,0 +1,7 @@
+const actionMesuresDelete = require("./hasura-actions.mesures.delete");
+const actionMesuresExport = require("./hasura-actions.mesures.export");
+
+module.exports = {
+  actionMesuresDelete,
+  actionMesuresExport,
+};
