+++ packages/api/src/routes/hasura-actions/mesure/hasura-actions.mesures.routes.js
@@ -1,17 +1,10 @@
 const express = require("express");
-const XLSX = require("xlsx");
 
+const { actionMesuresDelete } = require("./actions");
+const { actionMesuresExport } = require("./actions");
 const actionsMesuresImporter = require("./mesures-import/actionsMesuresImporter");
 const checkImportMesuresParameters = require("./hasura-actions.mesures-import.checker");
 const hasuraActionErrorHandler = require("../../../middlewares/hasura-error-handler");
-const getMesures = require("../../../services/getMesures");
-const { Tis } = require("../../../models/Tis");
-const { Mesure } = require("../../../models/Mesure");
-const { isEnAttente } = require("@emjpm/core");
-const { getEmailUserDatas } = require("../../../email/email-user-data");
-const {
-  cancelReservationEmail,
-} = require("../../../email/cancel-reservation-email");
 
 const router = express.Router();
 
@@ -39,97 +32,15 @@ router.post(
 // hasura action: `export_mesures_file`
 router.post(
   "/export",
-  async (req, res, next) => {
-    try {
-      const { userId, serviceId } = req.user;
-      const mesures = await getMesures({ serviceId, userId });
-      const exportedMesures = mesures.map((mesure) => {
-        const {
-          numero_rg,
-          numero_dossier,
-          date_nomination,
-          code_postal,
-          ville,
-          pays,
-          civilite,
-          annee_naissance,
-          lieu_vie,
-          nature_mesure,
-          champ_mesure,
-          tis,
-          cabinet,
-        } = mesure;
-        return {
-          annee_naissance,
-          cabinet,
-          champ_mesure,
-          civilite,
-          code_postal,
-          date_nomination,
-          lieu_vie,
-          nature_mesure,
-          numero_dossier,
-          numero_rg,
-          pays,
-          tribunal: tis ? tis.etablissement : "",
-          ville,
-        };
-      });
-      if (mesures.length) {
-        var ws = XLSX.utils.json_to_sheet(exportedMesures, {
-          header: Object.keys(exportedMesures[0]),
-        });
-        const wb = XLSX.utils.book_new();
-        XLSX.utils.book_append_sheet(wb, ws, "Mesures eMJPM");
-        res.header("Content-Type", "application/octet-stream");
-        res.attachment("export_mesures.xlsx");
-
-        var wopts = { bookSST: false, bookType: "xlsx", type: "base64" };
-        var wbout = XLSX.write(wb, wopts);
-
-        return res.send({ data: wbout });
-      } else {
-        return res.status(404).end();
-      }
-    } catch (err) {
-      return next(err);
-    }
-  },
+  actionMesuresExport,
   hasuraActionErrorHandler("Unexpected error processing file")
 );
 
 // hasura action: `delete_mesure`
-router.post("/delete", async function (req, res) {
-  const { mesure_id } = req.body.input;
-
-  try {
-    const mesure = await Mesure.query().findById(mesure_id);
-    const { ti_id, service_id, mandataire_id, status } = mesure;
-
-    if (!isEnAttente({ status }))
-      throw new Error(
-        `Delete needs a mesure with en_attente status (The mesure_id ${mesure.id} has ${status} status)`
-      );
-
-    const nbDeleted = await Mesure.query().deleteById(mesure_id);
-
-    if (nbDeleted === 0) return res.json({ success: false });
-
-    const ti = await Tis.query().findById(ti_id);
-    const users = await getEmailUserDatas(mandataire_id, service_id);
-    const emails = users.map((user) =>
-      cancelReservationEmail(ti, mesure, user)
-    );
-
-    Promise.all(emails);
-  } catch (error) {
-    console.error(error);
-    hasuraActionErrorHandler("Unexpected error with delete mesure");
-
-    return res.json({ success: false });
-  }
-
-  res.json({ success: true });
-});
+router.post(
+  "/delete",
+  actionMesuresDelete,
+  hasuraActionErrorHandler("Unexpected error deleting mesure")
+);
 
 module.exports = router;
